various improvements
mainly - dual emotion detection.

We now use emotion2vec+ and create a voice profile for the speaker for each emotion that is confirmed.

This overrides emotion2vec as it seems to be much more reliable, however there are rules that - of emotion2vec is neutral then the outcome with be neutral, in other cases the voice profile takes precedence.

This method provides much more neutrals and less false positives as you would expect normal conversations.

performance is negligible (on 3090) and real time recording through the WebSocket provides almost instant transcribed responses (along with the extra data).